# feat(perf): performance enhancements from span and generated regex