--- a/perf/PerformanceTests.cs
+++ b/perf/PerformanceTests.cs
@@ -5,33 +5,33 @@ namespace Tvans.Kenteken.PerformanceTests;
 [MemoryDiagnoser]
 public class PerformanceTests
 {
+    [Params(
+        "99-XX-XX", 
+        "99-99-XX",
+        "99-XX-99",
+        "XX-99-XX",
+        "XX-XX-99",
+        "99-XX-XX",
+        "99-XXX-9",
+        "9-XXX-99",
+        "XX-999-X",
+        "X-999-XX",
+        "XXX-99-x",
+        "X-99-XXX",
+        "9-XX-999",
+        "999-XX-9"
+    )]
+    public string Input { get; set; }
+    
     [Benchmark]
     public int? GetSidecode()
     {
-        return Kenteken.GetSidecode("55-GJ-GJ");
-    }
-    
-    [Benchmark]
-    public bool Validate()
-    {
-        return Kenteken.Validate("55-GJ-GJ");
-    }
-
-    [Benchmark]
-    public string Format()
-    {
-        return new Kenteken("55GJGJ").Formatted;
+        return Kenteken.GetSidecode(Input);
     }
     
     [Benchmark]
     public Kenteken New()
     {
-        return new Kenteken("55-GJ-GJ");
-    }
-    
-    [Benchmark]
-    public bool TryParse()
-    {
-        return Kenteken.TryParse("55-GJ-GJ", out _);
+        return new Kenteken(Input);
     }
 }