… README.md
… src/Tvans.Kenteken.csproj
… test/KentekenTests.cs