- feat(perf): use generated regex for better performance
- feat(perf): use Span and ReadOnlySpan where possible for better performance
- chore: drop support for .NET standard 2.0 and thus .NET Framework
- chore: bump version to 2.0.0 due to breaking change in dropping support for older frameworks

Improvements:

**Old**
| Method      | Input    | Mean       | Error    | StdDev   | Gen0   | Allocated |
|------------ |--------- |-----------:|---------:|---------:|-------:|----------:|
| GetSidecode | 99-99-XX |   972.6 ns | 11.63 ns | 10.31 ns | 0.0515 |     216 B |
| New         | 99-99-XX | 2,169.2 ns | 26.42 ns | 23.42 ns | 0.2937 |    1232 B |
| GetSidecode | 99-XX-99 | 1,343.3 ns | 23.49 ns | 20.82 ns | 0.0515 |     216 B |
| New         | 99-XX-99 | 2,570.5 ns | 14.35 ns | 11.98 ns | 0.2937 |    1232 B |
| GetSidecode | 99-XX-XX | 2,251.0 ns | 43.48 ns | 40.67 ns | 0.0496 |     216 B |
| New         | 99-XX-XX | 3,482.8 ns | 14.88 ns | 13.19 ns | 0.2937 |    1232 B |


**New**
| Method      | Input    | Mean       | Error    | StdDev    | Gen0   | Allocated |
|------------ |--------- |-----------:|---------:|----------:|-------:|----------:|
| GetSidecode | 99-99-XX |   413.1 ns |  4.90 ns |   4.58 ns |      - |         - |
| New         | 99-99-XX | 1,305.7 ns | 49.49 ns | 145.92 ns | 0.1831 |     768 B |
| GetSidecode | 99-XX-99 |   504.3 ns | 10.04 ns |  13.41 ns |      - |         - |
| New         | 99-XX-99 | 1,237.2 ns | 17.83 ns |  15.81 ns | 0.1831 |     768 B |
| GetSidecode | 99-XX-XX |   722.3 ns | 14.46 ns |  15.48 ns |      - |         - |
| New         | 99-XX-XX | 1,426.6 ns | 27.22 ns |  25.46 ns | 0.1831 |     768 B |